+++ integration_tests/test_dict_06.py
@@ -6,7 +6,7 @@ def fill_rollnumber2cpi(size: i32) -> dict[i32, f64]:
     
     rollnumber2cpi[0] = 1.1
     for i in range(1000, 1000 + size):
-        rollnumber2cpi[i] = float(i/100.0 + 5.0)
+        rollnumber2cpi[i] = float(i)/100.0 + 5.0
 
     return rollnumber2cpi
 
@@ -16,7 +16,7 @@ def test_dict():
     rollnumber2cpi: dict[i32, f64] = fill_rollnumber2cpi(size)
 
     for i in range(1000 + size - 1, 1001, -1):
-        assert abs(rollnumber2cpi[i] - i/100.0 - 5.0) <= 1e-12
+        assert abs(rollnumber2cpi[i] - f64(i)/100.0 - 5.0) <= 1e-12
 
     assert abs(rollnumber2cpi[0] - 1.1) <= 1e-12
     assert len(rollnumber2cpi) == 201
